+++ CHANGELOG.rst
@@ -1,6 +1,17 @@
 Changelog
 ---------
 
+0.7.1 (2018-08-07)
+~~~~~~~~~~~~~~~~~~
+* fixed gh #8 es_dangerous_reset --es-only to sync database to ES
+* fixed gh #17 make es_dangerous_reset remove dem models
+* improved test coverage
+* added tests for ``es_create --es-only``
+* added ``IndexVersion.hard_delete()`` (not called by default)
+* added ``hard_delete`` flag to ``DropIndexAction``
+* added ``hard_delete`` flag to ``DEMIndexManager.test_post_teardown()``
+* updated ``__str__()`` of ``IndexAction`` to be more descriptive
+
 0.7.0 (2018-08-06)
 ~~~~~~~~~~~~~~~~~~
 * fixed gh #5: "add python 3 support and tests"